fix: heading speed w3c spec (#270)
* CB-11462: (all) Make sure Coordinates#speed follows the w3c specification.

Speed can have a negative value on iOS devices when the OS is not able to determine the current speed.
This is against the w3c specification. Which only allows null and a number value >= 0.

See the following links for further informations:

 * https://dev.w3.org/geo/api/spec-source.html#speed
 * https://developer.apple.com/library/ios/documentation/CoreLocation/Reference/CLLocation_Class/#//apple_ref/occ/instp/CLLocation/speed

* CB-11462: (all) Make sure Coordinates#heading follows the w3c specification.

Heading (course) can have a negative value on iOS devices when the OS is not able to determine the current heading.
This is against the w3c specification. Which only allows null and a number value >= 0 and <= 360.

See the following links for further informations:

 * https://dev.w3.org/geo/api/spec-source.html#heading
 * https://developer.apple.com/library/ios/documentation/CoreLocation/Reference/CLLocation_Class/#//apple_ref/occ/instp/CLLocation/course

---------

Co-authored-by: Malte Legenhausen <mlegenhausen@gmail.com>